move change pass controller to user
